+++ packages/mobx/lib/src/api/extensions/observable_map_extension.dart
@@ -4,7 +4,4 @@ part of '../extensions.dart';
 extension ObservableMapExtension<K, V> on Map<K, V> {
   ObservableMap<K, V> asObservable({ReactiveContext? context, String? name}) =>
       ObservableMap<K, V>.of(this, context: context, name: name);
-
-  ObservableMap<K, V> obsIdentity({ReactiveContext? context, String? name}) =>
-      ObservableMap<K, V>.identity(this, context: context, name: name);
 }