+++ packages/mobx/lib/src/core/atom_extensions.dart
@@ -6,7 +6,14 @@ extension AtomSpyReporter on Atom {
     reportObserved();
   }
 
-  void reportWrite<T>(T newValue, T oldValue, void Function() setNewValue) {
+  void reportWrite<T>(T newValue, T oldValue, void Function() setNewValue, {EqualityComparer<T>? equals}) {
+    final areEqual = equals == null ? oldValue == newValue : equals(oldValue, newValue);
+
+    // Avoid unnecessary observable notifications of @observable fields of Stores
+    if (areEqual) {
+      return;
+    }
+
     context.spyReport(ObservableValueSpyEvent(this, newValue: newValue, oldValue: oldValue, name: name));
 
     final actionName = context.isSpyEnabled ? '${name}_set' : name;
@@ -14,7 +21,7 @@ extension AtomSpyReporter on Atom {
     // ignore: cascade_invocations
     context.conditionallyRunInAction(() {
       setNewValue();
-      reportChanged(areSpiesNotified: true);
+      reportChanged();
     }, this, name: actionName);
 
     // ignore: cascade_invocations