+++ packages/mobx/lib/src/core/context.dart
@@ -258,7 +258,8 @@ class ReactiveContext {
     return result;
   }
 
-  void _reportObserved(Atom atom) {
+  @protected
+  void reportObserved(Atom atom) {
     final derivation = _state.trackingDerivation;
 
     if (derivation != null) {
@@ -333,7 +334,9 @@ class ReactiveContext {
         _resetState();
 
         throw MobXCyclicReactionException(
-            "Reaction doesn't converge to a stable state after ${config.maxIterations} iterations. Probably there is a cycle in the reactive function: $failingReaction");
+            "Reaction doesn't converge to a stable state after ${config.maxIterations} iterations. "
+            "Probably there is a cycle in the reactive function: $failingReaction "
+            "(creation stack: ${failingReaction.debugCreationStack})");
       }
 
       final remainingReactions = allReactions.toList(growable: false);
@@ -395,7 +398,8 @@ class ReactiveContext {
     }
   }
 
-  void _clearObservables(Derivation derivation) {
+  @protected
+  void clearObservables(Derivation derivation) {
     final observables = derivation._observables;
     derivation._observables = {};
 
@@ -511,11 +515,13 @@ class ReactiveContext {
     _state.allowStateChanges = allow;
   }
 
-  void _pushComputation() {
+  @protected
+  void pushComputation() {
     _state.computationDepth++;
   }
 
-  void _popComputation() {
+  @protected
+  void popComputation() {
     _state.computationDepth--;
   }